requirements/requirements.txt
tornado==5.0.2
flask==1.0.2
flask-Cors==3.0.4
flask-httpauth==3.2.3
pyjwt==1.6.4
requests==2.18.4
glom==18.1.1
hvac==0.5.0
boto3==1.7.28
